Support Restify
The Ads team are using Restify for some of their services, this has a
very similar API to Express and only a few small tweaks are needed to
this library in order for it to work.

I think this is worth doing to get the library used across more of the
FT. I've done the appropriate work and we're also testing against
Restify to ensure it works in future.